+++ CHANGELOG.md
@@ -83,6 +83,7 @@ Added:
 - `printGlobalStateSCC` renamed to `printGlobalStateApp`.
 
 - The ` PyASCCache` has been merged to `ASCCache` and is not used anymore.
+- Only use list transaction in executeTx.
 
 ### Bug fixes